Restructure data_injector files
